Reorganized structure and modified Dockerfile
